chore: moved off topic code to off topic file
Let's keep the code for reference, but get it ready for deletion once we feel comfortable deleting it.
