+++ integration_test/off_topic_code.dart
@@ -7,13 +7,13 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart' hide Size;
 import 'package:ten_ten_one/bridge_definitions.dart';
-import 'package:ten_ten_one/main.dart' as app;
+import 'package:ten_ten_one/off_topic_code.dart' as app;
 import 'package:flutter_test/flutter_test.dart';
 import 'package:integration_test/integration_test.dart';
 import 'package:vm_service/vm_service_io.dart';
 
 void main() {
-  print('integration_test/main.dart starts!');
+  print('integration_test/off_topic_code.dart starts!');
 
   IntegrationTestWidgetsFlutterBinding.ensureInitialized();
