+++ lib/off_topic_code.dart
@@ -7,6 +7,41 @@ import 'dart:typed_data';
 
 import 'package:flutter/material.dart';
 import 'package:ten_ten_one/bridge_definitions.dart';
+import 'package:ten_ten_one/off_topic_code.dart';
+
+import 'ffi.io.dart' if (dart.library.html) 'ffi.web.dart';
+export 'ffi.io.dart' if (dart.library.html) 'ffi.web.dart' show api;
+
+void main() => runApp(const MyApp());
+
+class MyApp extends StatefulWidget {
+  const MyApp({Key? key}) : super(key: key);
+
+  @override
+  State<MyApp> createState() => _MyAppState();
+}
+
+class _MyAppState extends State<MyApp> {
+  Uint8List? exampleImage;
+  String? exampleText;
+
+  @override
+  void initState() {
+    super.initState();
+    _callExampleFfiTwo();
+  }
+
+  @override
+  Widget build(BuildContext context) => buildPageUi(
+        exampleImage,
+        exampleText,
+      );
+
+  Future<void> _callExampleFfiTwo() async {
+    final receivedText = await api.passingComplexStructs(root: createExampleTree());
+    if (mounted) setState(() => exampleText = receivedText);
+  }
+}
 
 Widget buildPageUi(Uint8List? exampleImage, String? exampleText) {
   return MaterialApp(
